fixing the python unit test dependencies
